python.test/requirements.txt
streamlit
python-docx
langchain
llama-index
openai